requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2.16
django-appconf==1.0.5
django-compressor==4.1
django-environ==0.9.0
django-libsass==0.9
django-phonenumber-field==7.0.0
django-sass-processor==1.2.1
libsass==0.21.0
rcssmin==1.1.0
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
autopep8==1.5
gunicorn==20.0.4
pycodestyle==2.5.0
pytz==2019.3
certifi==2022.9.24
charset-normalizer==2.1.1
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
requests==2.28.1
urllib3==1.26.12
yarg==0.1.9
psycopg2==2.9.4
python-dotenv==0.21.0
django-crontab==0.7.1
djangorestframework==3.14.0
